requirements.txt
pandas==2.2.3
openai==1.63.2
langchain==0.3.19 
langchain_openai==0.3.6
python-dotenv==1.0.1
langchain_chroma==0.2.2